rendum reorganisation db module
