+++ telebot/modules/db/exceptions.py
